ANLG-249: extract attachment sync core
