+++ Cargo.toml
@@ -55,6 +55,7 @@ anlg-apple-calendar = { path = "crates/apple-calendar", package = "apple-calenda
 anlg-apple-note = { path = "crates/apple-note", package = "apple-note" }
 anlg-apple-todo = { path = "crates/apple-todo", package = "apple-todo" }
 anlg-askama-utils = { path = "crates/askama-utils", package = "askama-utils" }
+anlg-attachment-sync-core = { path = "crates/attachment-sync-core", package = "attachment-sync-core" }
 anlg-audacity = { path = "crates/audacity", package = "audacity" }
 anlg-audio = { path = "crates/audio", package = "audio" }
 anlg-audio-actual = { path = "crates/audio-actual", package = "audio-actual" }
