+++ plugins/attachment-sync/Cargo.toml
@@ -11,11 +11,13 @@ description = ""
 tauri-plugin = { workspace = true, features = ["build"] }
 
 [dev-dependencies]
+sha2 = { workspace = true }
 specta-typescript = { workspace = true }
 tempfile = { workspace = true }
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 
 [dependencies]
+anlg-attachment-sync-core = { workspace = true }
 anlg-e2ee = { workspace = true }
 anlg-fs-sync-core = { workspace = true }
 
@@ -24,15 +26,10 @@ tauri-plugin-db = { workspace = true }
 tauri-plugin-settings = { workspace = true }
 tauri-specta = { workspace = true, features = ["derive", "typescript"] }
 
-futures-util = { workspace = true }
-reqwest = { workspace = true, features = ["stream"] }
 serde = { workspace = true }
-sha2 = { workspace = true }
 specta = { workspace = true }
 sqlx = { workspace = true, features = ["runtime-tokio", "sqlite", "macros"] }
 tempfile = { workspace = true }
-thiserror = { workspace = true }
 tokio = { workspace = true, features = ["fs", "io-util", "macros"] }
 tokio-util = { workspace = true }
-url = { workspace = true }
 uuid = { workspace = true, features = ["v4"] }
